@@ -35,17 +35,16 @@ repositories {
 apply plugin: 'com.android.library'
 
 ext {
-    versionName = '3.2.0'
+    versionName = '3.3.0'
 }
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.0"
+    compileSdkVersion 29
     publishNonDefault true
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 28
+        minSdkVersion 19
+        targetSdkVersion 29
     }
 
     sourceSets {
@@ -70,10 +69,10 @@ android {
 
 
 dependencies {
-    api "com.android.support:appcompat-v7:28.0.0-alpha3"
-    api "com.android.support:mediarouter-v7:28.0.0-alpha3"
-    api "com.android.support:support-v4:28.0.0-alpha3"
-    api 'com.google.android.gms:play-services-cast:16.0.0'
+    api 'androidx.appcompat:appcompat:1.0.2'
+    api 'androidx.mediarouter:mediarouter:1.0.0'
+    api 'androidx.legacy:legacy-support-v4:1.0.0'
+    api 'com.google.android.gms:play-services-cast:17.0.0'
     api 'com.github.bumptech.glide:glide:4.5.0'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.5.0'
 }
