@@ -1,7 +1,7 @@
 package com.google.android.libraries.cast.companionlibrary.cast;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
