@@ -21,7 +21,8 @@ import com.google.android.gms.common.ConnectionResult;
 import com.google.android.libraries.cast.companionlibrary.cast.BaseCastManager;
 import com.google.android.libraries.cast.companionlibrary.cast.exceptions.OnFailedListener;
 
-import android.support.v7.media.MediaRouter.RouteInfo;
+import androidx.mediarouter.media.MediaRouter.RouteInfo;
+import androidx.mediarouter.app.MediaRouteButton;
 
 /**
  * An interface for receiving callbacks around the connectivity status to a Cast device.
@@ -76,7 +77,7 @@ public interface BaseCastConsumer extends OnFailedListener {
     /**
      * Called when the number of cast devices present on the network changes from 0 to a positive
      * number or vice versa. Can be used, for example, to control the visibility of {@link
-     * android.support.v7.app.MediaRouteButton}
+     * MediaRouteButton}
      *
      * @param castPresent set to {@code true} if at least one device becomes available,
      * {@code false} otherwise
