@@ -23,7 +23,7 @@ import android.graphics.BitmapFactory;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v7.app.MediaRouteControllerDialog;
+import androidx.mediarouter.app.MediaRouteControllerDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.ImageView;
