@@ -26,10 +26,12 @@ import android.graphics.Point;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.DialogFragment;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+
 import android.text.TextUtils;
 import android.view.View;
 import android.widget.SeekBar;
@@ -648,7 +650,7 @@ public class VideoCastControllerFragment extends Fragment implements OnVideoCast
      * Shows an error dialog
      */
     private void showErrorDialog(String message) {
-        android.support.v4.app.FragmentManager fragmentManager = getFragmentManager();
+        FragmentManager fragmentManager = getFragmentManager();
         if (fragmentManager != null)
             ErrorDialogFragment.newInstance(message).show(fragmentManager, "dlg");
     }
