@@ -18,7 +18,7 @@ package com.google.android.libraries.cast.companionlibrary.cast.tracks.ui;
 
 import android.app.Activity;
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
