@@ -1,7 +1,7 @@
 package com.google.android.libraries.cast.companionlibrary.cast.ui;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.android.libraries.cast.companionlibrary.cast.CastManagerBuilder;
 
