@@ -34,10 +34,10 @@ import android.graphics.PorterDuffXfermode;
 import android.graphics.Rect;
 import android.os.Build;
 import android.preference.PreferenceManager;
-import android.support.annotation.ColorRes;
-import android.support.annotation.DimenRes;
-import android.support.annotation.StringRes;
-import android.support.v7.app.MediaRouteButton;
+import androidx.annotation.ColorRes;
+import androidx.annotation.DimenRes;
+import androidx.annotation.StringRes;
+import androidx.mediarouter.app.MediaRouteButton;
 import android.text.TextUtils;
 import android.util.AttributeSet;
 import android.view.LayoutInflater;
